ai/requirements.txt
# Python dependencies for AI service
fastapi
uvicorn
chromadb
sentence-transformers
pydantic
requests
python-dotenv
google-generativeai>=0.3.0